Moved reperesentations and request parameters from Service module
