+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/ServiceStatus.java
@@ -7,7 +7,7 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common;
 
 public enum ServiceStatus {
   ACTIVE,
