+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/ParameterDefinitionRepresentation.java
@@ -7,13 +7,14 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.types.enumtypes.ParameterModificationType;
 import org.oscm.internal.types.enumtypes.ParameterType;
 import org.oscm.internal.types.enumtypes.ParameterValueType;
 import org.oscm.internal.vo.VOParameterDefinition;
 import org.oscm.internal.vo.VOParameterOption;
+import org.oscm.rest.common.representation.ParameterOptionRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
