+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/ParameterOptionRepresentation.java
@@ -7,7 +7,7 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.vo.VOParameterOption;
 import org.oscm.rest.common.representation.Representation;
