+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/ParameterRepresentation.java
@@ -7,9 +7,10 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.vo.VOParameter;
+import org.oscm.rest.common.representation.ParameterDefinitionRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
