+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/PricedEventRepresentation.java
@@ -7,10 +7,9 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.vo.VOPricedEvent;
-import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
 import java.math.BigDecimal;
