+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/TechnicalServiceRepresentation.java
@@ -7,11 +7,10 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.service.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.types.enumtypes.ServiceAccessType;
 import org.oscm.internal.vo.*;
-import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
 import java.util.ArrayList;
