+++ oscm-rest-api-service/src/main/java/org/oscm/rest/service/CompatibleServiceResource.java
@@ -15,7 +15,8 @@ import org.oscm.rest.common.CommonParams;
 import org.oscm.rest.common.representation.RepresentationCollection;
 import org.oscm.rest.common.RestResource;
 import org.oscm.rest.common.Since;
-import org.oscm.rest.service.data.ServiceRepresentation;
+import org.oscm.rest.common.requestparameters.ServiceParameters;
+import org.oscm.rest.common.representation.ServiceRepresentation;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
