+++ oscm-rest-api-service/src/main/java/org/oscm/rest/service/PriceModelResource.java
@@ -14,7 +14,8 @@ import lombok.Setter;
 import org.oscm.rest.common.CommonParams;
 import org.oscm.rest.common.RestResource;
 import org.oscm.rest.common.Since;
-import org.oscm.rest.service.data.PriceModelRepresentation;
+import org.oscm.rest.common.requestparameters.ServiceParameters;
+import org.oscm.rest.common.representation.PriceModelRepresentation;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
