+++ oscm-rest-api-service/src/main/java/org/oscm/rest/service/ServiceBackend.java
@@ -16,9 +16,10 @@ import org.oscm.internal.vo.VOService;
 import org.oscm.internal.vo.VOServiceDetails;
 import org.oscm.rest.common.representation.RepresentationCollection;
 import org.oscm.rest.common.RestBackend;
-import org.oscm.rest.service.data.ServiceDetailsRepresentation;
-import org.oscm.rest.service.data.ServiceRepresentation;
-import org.oscm.rest.service.data.StatusRepresentation;
+import org.oscm.rest.common.requestparameters.ServiceParameters;
+import org.oscm.rest.common.representation.ServiceDetailsRepresentation;
+import org.oscm.rest.common.representation.ServiceRepresentation;
+import org.oscm.rest.common.representation.StatusRepresentation;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
