+++ oscm-rest-api-service/src/main/java/org/oscm/rest/service/TechnicalServiceResource.java
@@ -17,7 +17,8 @@ import org.oscm.internal.vo.VOTechnicalService;
 import org.oscm.rest.common.CommonParams;
 import org.oscm.rest.common.RestResource;
 import org.oscm.rest.common.Since;
-import org.oscm.rest.service.data.TechnicalServiceRepresentation;
+import org.oscm.rest.common.requestparameters.ServiceParameters;
+import org.oscm.rest.common.representation.TechnicalServiceRepresentation;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
