+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/ServiceBackendTest.java
@@ -13,9 +13,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.oscm.internal.intf.ServiceProvisioningService;
 import org.oscm.internal.vo.VOServiceDetails;
 import org.oscm.internal.vo.VOTechnicalService;
-import org.oscm.rest.common.representation.RepresentationCollection;
+import org.oscm.rest.common.ServiceStatus;
+import org.oscm.rest.common.representation.*;
 import org.oscm.rest.common.SampleTestDataUtility;
-import org.oscm.rest.service.data.*;
+import org.oscm.rest.common.requestparameters.ServiceParameters;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
