+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/EventDefinitionRepresentationTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.oscm.internal.types.enumtypes.EventType;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOEventDefinition;
+import org.oscm.rest.common.representation.EventDefinitionRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import static org.assertj.core.api.Assertions.assertThat;
