+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/OperationRepresentationTest.java
@@ -13,6 +13,8 @@ import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOServiceOperationParameter;
 import org.oscm.internal.vo.VOTechnicalServiceOperation;
+import org.oscm.rest.common.representation.OperationParameterRepresentation;
+import org.oscm.rest.common.representation.OperationRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import java.util.ArrayList;
