+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/ParameterOptionRepresentationTest.java
@@ -12,6 +12,7 @@ package org.oscm.rest.service.data;
 import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOParameterOption;
+import org.oscm.rest.common.representation.ParameterOptionRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import static org.assertj.core.api.Assertions.assertThat;
