+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/ParameterRepresentationTest.java
@@ -13,6 +13,8 @@ import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOParameter;
 import org.oscm.internal.vo.VOParameterDefinition;
+import org.oscm.rest.common.representation.ParameterDefinitionRepresentation;
+import org.oscm.rest.common.representation.ParameterRepresentation;
 import org.oscm.rest.common.representation.Representation;
 
 import static org.assertj.core.api.Assertions.assertThat;
