+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/PricedParameterRepresentationTest.java
@@ -11,6 +11,7 @@ package org.oscm.rest.service.data;
 
 import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.*;
+import org.oscm.rest.common.representation.*;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
