+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/RoleDefinitionRepresentationTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VORoleDefinition;
 import org.oscm.rest.common.representation.Representation;
+import org.oscm.rest.common.representation.RoleDefinitionRepresentation;
 
 import static org.assertj.core.api.Assertions.assertThat;
