+++ oscm-rest-api-service/src/test/java/org/oscm/rest/service/data/SteppedPriceRepresentationTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOSteppedPrice;
 import org.oscm.rest.common.representation.Representation;
+import org.oscm.rest.common.representation.SteppedPriceRepresentation;
 
 import java.math.BigDecimal;
